Add docs, move libpkl to internal pkg
